+++ src/Cart/Contracts/CartItem.php
@@ -18,4 +18,7 @@ use Vanilo\Contracts\CheckoutSubjectItem;
 
 interface CartItem extends CheckoutSubjectItem
 {
+    public function hasParent(): bool;
+
+    public function getParent(): ?self;
 }