+++ Changelog.md
@@ -9,6 +9,8 @@
 - Changed the minimum Laravel 10 version to v10.48
 - Added Laravel 12 Support
 - BC: Added the `getBuyable()` method to the `OrderItem` interface
+- BC: Added the `addSubItem()` method to the Cart interface
+- BC: Added the `hasParent()` and `getParent()` methods to the CartItem interface
 - Added the `gtin` field to the products and master product variants tables
 - Fixed the missing float cast of the order item's price field
 - PHP 8.4 deprecation notice fixes
@@ -25,6 +27,7 @@
 - Changed the `Order` and `OrderItem` interfaces to extend the new `Sale` and `SaleItem` interfaces from the contracts module
 - Fixed issue where the `linkable` pointed to a missing product thus returning null
 - Added the `excludingSkus()` method to `ProductSearch`
+- Added SubItem support to the cart items
 
 ## 4.x Series