+++ src/Cart/Changelog.md
@@ -1,5 +1,17 @@
 # Vanilo Cart Module Changelog
 
+## 5.x Series
+
+## Unreleased
+##### 2025-XX-XX
+
+- BC: Added the `addSubItem()` method to the Cart interface
+- BC: Added the `hasParent()` and `getParent()` methods to the CartItem interface
+- Dropped PHP 8.2 Support
+- Changed the minimum Laravel 10 version to v10.48
+- Added Laravel 12 Support
+- Added SubItem support to the cart items
+
 ## 4.x Series
 
 ## 4.2.0